replaced portfolioClient by operationsService

+++ src/main/java/ru/obukhov/trader/market/model/PlacedMarketOrder.java
@@ -3,14 +3,16 @@ package ru.obukhov.trader.market.model;
 
 import ru.tinkoff.piapi.contract.v1.OperationType;
 
+import java.math.BigDecimal;
+
 public record PlacedMarketOrder(
         String orderId,
         OperationType operation,
         OrderStatus status,
         String rejectReason,
         String message,
-        Long requestedLots,
-        Long executedLots,
+        BigDecimal requestedLots,
+        BigDecimal executedLots,
         MoneyAmount commission
 ) {
 }
